Add new mining sorting algorithm for heap
The old sorting algorithm would discriminate based on fees per KB for
tickets, even when tickets were both very small.  Because of the
limitations for numbers of tickets that may be included in the block,
along with the limited size of the block, it would be wiser when the
tickets are small to simply sort based on the absolute fee.

The option was also added in the configuration to allow users to set
their own mining algorithms as well as their own priority ticket size.
The new algorithm is disabled by default, and comments around caveats
for enabling it are given in the codebase.

fix typos and tests